backend/requirements.txt
fastapi
uvicorn
wsproto
sarvamai
python-dotenv
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client
slack_sdk
